[DotNetCore] Allow .NET Core location to be configured
Fixes VSTS #566365 Support .NET Core SDK being installed in
non-standard location

In Preferences there is now a Projects - SDK Locations - .NET Core
section that can be used to configure the location of the .NET Core
command line tool (dotnet). This can be used to point to a different
non-standard install location of the .NET Core SDK. After this is
switched the MSBuild hosts are recycled and all .NET Core projects
are re-evaluated to ensure the new locations of any MSBuild targets
is used.