[Ide] Add an internal method to allow reloading extensions attached to a document
Context: https://github.com/mono/monodevelop/issues/3870

This commits adds an internal method that we can call to reload the extension chain and allow the pathbar widget to be updated when a view content changes the underlying text editor instances it uses (and thus alter the value that can be returned from GetContent).

Note that it's likely still profitable to add a ContentChanged event of some sort in the future to handle the use case described in the issue but that can be done on a longer term effort to support those. In the meantime that method is an acceptable workaround for the subset of features we care about.