[Core] Fix duplicate key restore error if duplicate files in project
If a .NET Core project has a duplicate MSBuild item or an MSBuild
item that is already implicitly added then the NuGet restore would
fail with an error similar to:

An item with the same key has already been added. Key: (None, dupe.txt)

Now duplicate MSBuild items are handled when the project is
re-evaluated.