[DotNetCore] Fix dotnet cli file selector dialog cancel behaviour
In the .NET Core SDK preferences if the file selector dialog that
is used to browse to the dotnet command line is cancelled, then
on clicking the browse button again would not reset the folder to
that set in the text box. Instead the previously selected folder
would be used. This was inconsistent with the behaviour of the other
SDKs panels in preferences.